Create utils folder and split utils.js in multiple files
